pyproject.toml
[tool.poetry]
name = "python-cdk-project-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.2"
"aws-cdk.core" = "^1.119.0"
black = "^21.7b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
python-dotenv = "^0.19.0"
pytest = "^6.2.4"
"aws-cdk.aws-s3" = "^1.119.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
